Reorganizing tests, adding some more docs
