+++ astroquery/exoplanets/__init__.py
@@ -5,3 +5,5 @@ from __future__ import (absolute_import, division, print_function,
 __all__ = ['PlanetParams']
 
 from .planet_params import PlanetParams
+
+
